New regression test results as ce5091de4cd8f92bb3b304841cd9b8b9d48d500c fixed handover and that amounts to faster handover times now
scalars don't delete stats.csv files for ConfTest test

+++ Etc/Scalars/scalars.rb
@@ -329,7 +329,9 @@ end
       end   
       wait = `cut -d ',' -f 1-4 "#{e}.mean.csv" | paste -d "," - #{stats.join(".tmp ") + ".tmp"} > "#{e}.csv"`
       wait = `rm #{stats.join(".tmp ") + ".tmp"}`
-      stats.each{|i| `rm "#{e}.#{i}.csv"` } 
+      if not @debug
+        stats.each{|i| `rm "#{e}.#{i}.csv"` } 
+      end
     }
     array
   end
